Complete flex exercise 05
I took a different route than the solution suggests. I moved the X
button outside of the div that I created around the main content. I was
then able to line it up using the modal flex container instead of
applying an additional flex container to the header like the solution
suggests. I'm not sure which way is cleaner but my solution makes sense
to me and gets the desired result.